backend/requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
PyPDF2
beautifulsoup4
httpx
python-multipart
pydantic-settings
python-dotenv
groq
google-generativeai
slowapi
